Added the timesquestion program
The timesquestion program demonstrates selection in go.
The timesquestion program asks the pupil a random multiplication
from the 1..12 times tables. If the pupil enters the correct answer
the program says "Congratulations". If the pupil enters a wrong answer
the program tells them if their answer was to large or to small and also
tells them what the correct answer was.